refactor all files Relates # 31
