+++ public/index.html
@@ -16,6 +16,8 @@
 
     </div>
     <script src="../public/js/logic.js" type="application/javascript" ></script>
+    <script src="../public/js/dom.js" type="application/javascript" ></script>
+
     <script src="../public/js/request.js" type="text/javascript"></script>
 </body>
 </html>
