+++ test/test.js
@@ -1,5 +1,5 @@
 const test = require('tape');
-const autocomplete = require('../public/js/logic')
+const getResultArr = require('../public/js/logic')
 const countriesArr = [
     {
         "name": "Afghanistan",
@@ -40,7 +40,7 @@ const countriesArr = [
 ]
 
 test('Autocomplete test function', (t)=>{
-    const actual = autocomplete(countriesArr, 'a');
+    const actual = getResultArr(countriesArr, 'A');
     const expected = ['Afghanistan', 'Albania', 'Algeria'];
     t.deepEqual(actual, expected, '');
     t.end();
